Add option to trim leading underscore
Keys with unencrypted values have a leading underscore. When exported to
an env variable the underscore is preserved but can feel awkward to use
on the system.

The new `--trim-underscore` flag is introduced and can be used to, you
guessed it, trim the underscore from the exported env variable name.

The use of this flag makes `ejson2env` match Krane's
[behaviour](https://github.com/shopify/krane?tab=readme-ov-file#deploying-kubernetes-secrets-from-ejson)
of stripping the leading understore from secrets.

Fixes #17